Refactor: __tests__ 폴더 구조 변경
__tests__라는 한 폴더에 다 모여져있던 파일들을 src폴더처럼 mvc로 나눠 재배치했음
